fix(beamr): fix JIT exception handling compilation and state preservation
Three fixes for B-137 exception handling:
1. Set terminated=false after Try instruction — Try emits no Cranelift
   terminator, so the block boundary jump must be emitted by the compile
   loop (same pattern as the PutTuple2/PutList allocation fix).
2. Cast X_REGISTER_COUNT to usize in exception test register allocation.
3. Save/restore current_module and code_position in jit_call_interpreted
   so compiled frame metadata survives interpreter exception exits that
   clear module state.

+++ crates/beamr/src/jit/compiler/compiler_tests.rs
@@ -906,7 +906,7 @@ fn compiled_try_catches_interpreted_exception_and_exposes_payload() {
         caller.as_ref() as *const Module,
         &registry as *const ModuleRegistry,
     )));
-    let mut registers = vec![0; X_REGISTER_COUNT + 3];
+    let mut registers = vec![0; X_REGISTER_COUNT as usize + 3];
 
     let returned = call_native_status(&native, &mut registers, &mut process);
